+++ client/src/main.tsx
@@ -10,6 +10,7 @@ import PrayerTimesPage from "./pages/PrayerTimesPage";
 import IqamaTimesPage from "./pages/IqamaTimesPage";
 import JummahTimesPage from "./pages/JummahTimesPage";
 import AnnouncementsPage from "./pages/AnnouncementsPage";
+import EventsPage from "./pages/EventsPage";
 import NotificationsPage from "./pages/NotificationsPage";
 import PushRegistrationsPage from "./pages/PushRegistrationsPage";
 
@@ -49,6 +50,7 @@ ReactDOM.createRoot(document.getElementById("root")!).render(
             <Route path="jummah-times" element={<JummahTimesPage />} />
             <Route path="halal-places" element={<ResourcePage resource="halal-places" />} />
             <Route path="announcements" element={<AnnouncementsPage />} />
+            <Route path="events" element={<EventsPage />} />
             <Route path="notifications" element={<NotificationsPage />} />
             <Route path="quiz-questions" element={<ResourcePage resource="quiz-questions" />} />
             <Route path="push-registrations" element={<PushRegistrationsPage />} />